# src/swagger.json
{
	"components": {
...
		"schemas": {
			"ProfileType": {
				"type": "string",
				"enum": [
					"stakeholder",
					"admin",
					"unit owner"
				]
			},
			"RecordStatus": {
				"type": "string",
				"enum": [
					"active",
					"inactive"
				]
			},
			"AuthProfile": {
				"properties": {
					"remarks": {
						"type": "string"
					},
					"scopes": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/RecordStatus"
					},
					"type": {
						"$ref": "#/components/schemas/ProfileType"
					},
					"mobileNumber": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"status",
					"type",
					"email",
					"username",
					"name",
					"id"
				],
				"type": "object"
			},
			"AuthResult": {
				"properties": {
					"token": {
						"type": "string"
					},
					"profile": {
						"$ref": "#/components/schemas/AuthProfile"
					}
				},
				"required": [
					"token",
					"profile"
				],
				"type": "object"
			},
			"CommunityAttr": {
				"properties": {
					"description": {
						"type": "string"
					}
				},
				"required": [
					"description"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ChargeType": {
				"type": "string",
				"enum": [
					"unit",
					"percentage",
					"amount"
				]
			},
			"PostingType": {
				"type": "string",
				"enum": [
					"monthly",
					"manual",
					"accrued",
					"interest",
					"quarterly"
				]
			},
			"ChargeAttr": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"communityId": {
						"type": "number",
						"format": "double"
					},
					"community": {
						"$ref": "#/components/schemas/CommunityAttr"
					},
					"code": {
						"type": "string"
					},
					"rate": {
						"type": "number",
						"format": "double"
					},
					"chargeType": {
						"$ref": "#/components/schemas/ChargeType"
					},
					"postingType": {
						"$ref": "#/components/schemas/PostingType"
					},
					"thresholdInMonths": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"priority": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"passOn": {
						"type": "boolean",
						"nullable": true
					}
				},
				"required": [
					"communityId",
					"code",
					"rate",
					"chargeType",
					"postingType"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"TransactionType": {
				"type": "string",
				"enum": [
					"charged",
					"collected"
				]
			},
...
			"DisbursementBreakdownView": {
				"properties": {
					"passOn": {
						"type": "boolean"
					},
					"amount": {
						"type": "number",
						"format": "double"
					},
					"code": {
						"type": "string"
					}
				},
				"required": [
					"amount",
					"code"
				],
				"type": "object"
			},
			"PaymentType": {
				"type": "string",
				"enum": [
					"cash",
					"check"
				]
			},
			"ProfileAttr": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"mobileNumber": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/ProfileType"
					},
					"status": {
						"$ref": "#/components/schemas/RecordStatus"
					},
					"scopes": {
						"type": "string"
					},
					"remarks": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"username",
					"password",
					"email",
					"type",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DisbursementAttr": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"voucherId": {
						"type": "number",
						"format": "double"
					},
					"purchaseOrderId": {
						"type": "number",
						"format": "double"
					},
					"chargeId": {
						"type": "number",
						"format": "double"
					},
					"releasedBy": {
						"type": "number",
						"format": "double"
					},
					"paymentType": {
						"$ref": "#/components/schemas/PaymentType"
					},
					"details": {
						"type": "string"
					},
					"checkNumber": {
						"type": "string"
					},
					"checkPostingDate": {
						"type": "string",
						"format": "date-time"
					},
					"checkIssuingBank": {
						"type": "string"
					},
					"amount": {
						"type": "number",
						"format": "double"
					},
					"releasedByProfile": {
						"$ref": "#/components/schemas/ProfileAttr"
					},
					"charge": {
						"$ref": "#/components/schemas/ChargeAttr"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"chargeId",
					"releasedBy",
					"paymentType",
					"details",
					"amount"
				],
				"type": "object",
				"additionalProperties": false
			},
			"FieldError": {
				"properties": {
					"value": {
						"type": "string",
						"nullable": true
					},
					"field": {
						"type": "string",
						"nullable": true
					},
					"type": {
						"type": "string",
						"nullable": true
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"value",
					"field",
					"type",
					"message"
				],
				"type": "object"
			},
			"EntityError": {
				"properties": {
					"name": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					},
					"status": {
						"type": "number",
						"format": "double"
					},
					"fieldErrors": {
						"items": {
							"$ref": "#/components/schemas/FieldError"
						},
						"type": "array",
						"default": []
					}
				},
				"required": [
					"name",
					"message",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RegisterProfile": {
				"properties": {
					"mobileNumber": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password",
					"username",
					"name"
				],
				"type": "object"
			},
			"ApiError": {
				"properties": {
					"name": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"message",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateProfile": {
				"properties": {
					"remarks": {
						"type": "string",
						"nullable": true
					},
					"scopes": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/RecordStatus"
					},
					"type": {
						"$ref": "#/components/schemas/ProfileType"
					},
					"mobileNumber": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"status",
					"type",
					"email",
					"name"
				],
				"type": "object"
			},
			"PropertyAttr": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"code": {
						"type": "string"
					},
					"floorArea": {
						"type": "number",
						"format": "double"
					},
					"address": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/RecordStatus"
					}
				},
				"required": [
					"code",
					"floorArea",
					"address",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PaymentDetailAttr": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"collectedBy": {
						"type": "number",
						"format": "double"
					},
					"orNumber": {
						"type": "string"
					},
					"paymentType": {
						"$ref": "#/components/schemas/PaymentType"
					},
					"checkNumber": {
						"type": "string"
					},
					"checkPostingDate": {
						"type": "string",
						"format": "date-time"
					},
					"checkIssuingBank": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"collectedBy",
					"orNumber",
					"paymentType"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TransactionAttr": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"chargeId": {
						"type": "number",
						"format": "double"
					},
					"charge": {
						"$ref": "#/components/schemas/ChargeAttr"
					},
					"propertyId": {
						"type": "number",
						"format": "double"
					},
					"property": {
						"$ref": "#/components/schemas/PropertyAttr"
					},
					"amount": {
						"type": "number",
						"format": "double"
					},
					"transactionPeriod": {
						"type": "string",
						"format": "date-time"
					},
					"transactionType": {
						"$ref": "#/components/schemas/TransactionType"
					},
					"waivedBy": {
						"type": "number",
						"format": "double"
					},
					"comments": {
						"type": "string"
					},
					"paymentDetailId": {
						"type": "number",
						"format": "double"
					},
					"paymentDetail": {
						"$ref": "#/components/schemas/PaymentDetailAttr"
					},
					"rateSnapshot": {
						"type": "number",
						"format": "double"
					},
					"batchId": {
						"type": "string",
						"nullable": true
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"chargeId",
					"propertyId",
					"amount",
					"transactionPeriod",
					"transactionType"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PropertyAccount": {
				"properties": {
					"balance": {
						"type": "number",
						"format": "double"
					},
					"propertyId": {
						"type": "number",
						"format": "double"
					},
					"property": {
						"$ref": "#/components/schemas/PropertyAttr"
					},
					"assignedProfiles": {
						"items": {
							"$ref": "#/components/schemas/ProfileAttr"
						},
						"type": "array"
					},
					"transactions": {
						"items": {
							"$ref": "#/components/schemas/TransactionAttr"
						},
						"type": "array"
					},
					"paymentDetails": {
						"items": {
							"$ref": "#/components/schemas/PaymentDetailAttr"
						},
						"type": "array"
					}
				},
				"required": [
					"balance",
					"propertyId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Month": {
				"type": "string",
				"enum": [
					"JAN",
					"FEB",
					"MAR",
					"APR",
					"MAY",
					"JUN",
					"JUL",
					"AUG",
					"SEP",
					"OCT",
					"NOV",
					"DEC"
				]
			},
...
		},
		"securitySchemes": {
			"bearer": {
				"type": "apiKey",
				"in": "header",
				"name": "authorization"
			}
		}
	},
	"info": {
		"title": "luminous-api",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/api/disbursement/getDisbursementBreakdown": {
			"get": {
				"operationId": "GetDisbursementBreakdown",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/DisbursementBreakdownView"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [
					{
						"cookie": []
					}
				],
				"parameters": []
			}
		},
		"/api/disbursement/getAllDisbursements/{chargeId}": {
			"get": {
				"operationId": "GetAllDisbursements",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/DisbursementAttr"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [
					{
						"cookie": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "chargeId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "year",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/disbursement/postChargeDisbursement": {
			"post": {
				"operationId": "PostChargeDisbursement",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"security": [
					{
						"cookie": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DisbursementAttr"
							}
						}
					}
				}
			}
		},
		"/api/profile/getAll": {
			"get": {
				"operationId": "GetAllProfiles",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/AuthProfile"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [
					{
						"cookie": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/profile/register": {
			"post": {
				"operationId": "Register",
				"responses": {
					"201": {
						"description": "created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthResult"
								}
							}
						}
					},
					"400": {
						"description": "bad request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/EntityError"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RegisterProfile"
							}
						}
					}
				}
			}
		},
		"/api/profile/me": {
			"get": {
				"operationId": "Me",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthProfile"
								}
							}
						}
					}
				},
				"security": [
					{
						"cookie": []
					}
				],
				"parameters": []
			}
		},
		"/api/profile/updateProfileStatus/{id}": {
			"patch": {
				"operationId": "UpdateProfileStatus",
				"responses": {
					"204": {
						"description": "no content"
					}
				},
				"security": [
					{
						"cookie": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/RecordStatus"
						}
					}
				]
			}
		},
		"/api/profile/updateProfile/{id}": {
			"patch": {
				"operationId": "UpdateProfile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthProfile"
								}
							}
						}
					},
					"400": {
						"description": "bad request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/EntityError"
								}
							}
						}
					},
					"404": {
						"description": "not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				},
				"security": [
					{
						"cookie": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateProfile"
							}
						}
					}
				}
			}
		},
		"/api/profile/changePassword/{id}": {
			"patch": {
				"operationId": "ChangePassword",
				"responses": {
					"204": {
						"description": "no content"
					}
				},
				"security": [
					{
						"cookie": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"newPassword": {
										"type": "string"
									},
									"currentPassword": {
										"type": "string"
									}
								},
								"required": [
									"newPassword",
									"currentPassword"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/profile/resetPassword": {
			"post": {
				"operationId": "ResetPassword",
				"responses": {
					"204": {
						"description": "no content"
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"email": {
										"type": "string"
									},
									"username": {
										"type": "string"
									}
								},
								"required": [
									"email",
									"username"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/property-account/getPropertyAccountsByProfile/{profileId}": {
			"get": {
				"operationId": "GetPropertyAccountsByProfile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/PropertyAccount"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [
					{
						"cookie": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "profileId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/property-account/getPropertyAccount/{propertyId}": {
			"get": {
				"operationId": "GetPropertyAccount",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PropertyAccount"
								}
							}
						}
					}
				},
				"security": [
					{
						"cookie": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "propertyId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "year",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "month",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/Month"
						}
					}
				]
			}
		},
		"/api/property-account/getPropertyAccountsByPeriod": {
			"get": {
				"operationId": "GetPropertyAccountsByPeriod",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/PropertyAccount"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [
					{
						"cookie": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "year",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "month",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/Month"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}